requirements.txt
gradio
numpy
pillow
tflite-runtime==2.14.0
scikit-learn

